Don't keep the notification mutex in ty_device_monitor_refresh (win32)
The notification mutex was previously kept locked until all events
were processed, which may take time (due to callbacks). This could
keep the monitoring thread from processing WM_DEVICECHANGE which
MSDN says must be processed quickly.

Now, the notification list is moved to a local variable and the
mutex is released immediately. It is locked a second time after
processing the notifications to prepend any remaining event for
future processing (in case of an error).