Fix potential use-after-free in ty_board_wait_for
Manual refcounting is messy. In this case, in case the library user does
not keep a reference to the board it uses with ty_board_wait_for, things
could go wrong when/if the board itself is dropped.

The user code should not do this in the first place, but you never know.