Use select for ty_poll on OS X
poll uses kqueue in Darwin, which is broken for many types of
descriptors, such as tty and kqueue itself. Which won't work with
the upcoming darwin device implementation, since it needs to poll
mach ports though kqueue.

And of course, the monitor has to poll stdin, a tty in most cases.