Fix blocking behavior of ty_serial_write() on POSIX platforms
ty's documentation (in my head, that is) states that the device write
functions are always blocking, the block argument for ty_device_open()
affects only the read functions.

ty_serial_write() in device_posix.c did not obey that and this caused the
monitor to drop stuff when sending a lot of data to the board.