Allow asynchronous tasks to return an error message
This commit replaces the use of QFuture<void> with QFuture<QString>.
A finished task is considered a failure if there is no result
reported, or the first result is a non-empty string (error message).

The task return value is not correctly reported for now, the
next commit will implement the necessary changes on BoardTask.