Add parallel mode to tyb_monitor
tyb_monitor can be used in normal or parallel mode. In parallel mode, the wait
functions, such as tyb_monitor_wait(), do not refresh the monitor but wait for
someone else to do it (parallel thread)

This is used by TyQt to do work (upload, reset, reboot) in background threads.