Fix IRP leak in device read functions on Win32
This bug is bad, and I never noticed it because the leak goes
to non-paged kernel memory.

It was introduced in ty and libhs when fixing another bug where
the read functions would suddenly stall because they failed to
issue a new asynchronous ReadFile:
- Koromix/ty@ba44d72582ff83d556d8615e34b66d372aead365
- Koromix/libhs@0141622aae27c406bde3991e2deef13d78681d51

The problem is that _hs_win32_finalize_read() returning 0 could
mean that the previous read was still pending or that it finished
and returned 0 bytes.