libhs: Fix compilation with MSVC toolset v140_xp
For some reason some headers (hidsdi.h, etc.) are not available
with the XP-compatible toolset v140_xp. So we search for them
with FindWindowsSDK.cmake.

Honestly, I don't really get the whole SDK/WDK mess. Some headers
seem to be in both, some only in one. I don't get it, I don't
even know if this bug is caused by CMake or not.