Fix TyQt crash on Win32 due to double delete
Qt's version of WinMain in qtmain_win.cpp creates argv on the
heap and then deletes each pointer.

The C standard allows you do change the content of argv, but the
C++ standard does not say much about that. So I think the Qt code
is correct and my assumptions were not.