Reorganize source and resources files (except libhs)
Drop the separation between src and include for libraries, which
is annoying and pointless.

Drop libtyqt to simplify the build system. Instead we keep all the
common Qt files in src/tyqt and use them directly from src/upty.
That means some files are built twice, but it does not seem to
increase the compilation time, and it makes things simpler.

With that change we now have only one Qt resource file, which we
can move to resources/.